[CORRECTION] Évite les multiples duplications non désirées
… qui arrivaient en « ouvrant » puis « fermant » plusieurs fois de suite une modale de duplication.

Le souci était que le handler `.on('submit')` était déclaré **dans** le handler `.on('modaleDuplication')`.
Autrement dit : chaque ouverture de modale enregistrait un nouveau handler de soumission du formulaire.

Ce commit corrige le bug, et au passage introduit un fichier JS dédié à la modale pour mieux s'y retrouver.
