+++ public/espacePersonnel.js
@@ -3,9 +3,6 @@ import { brancheModale, initialiseComportementModale } from './modules/interacti
 import brancheComportementPastilles from './modules/interactions/pastilles.js';
 import brancheComportementSaisieContributeur from './modules/interactions/saisieContributeur.js';
 import brancheMenuContextuelService from './modules/interactions/brancheMenuContextuelService.js';
-import { brancheValidation, declencheValidation } from './modules/interactions/validation.mjs';
-
-const tableauDeLongueur = (longueur) => [...Array(longueur).keys()];
 
 $(() => {
   const $modaleSuppression = $('.modale-suppression-service');
@@ -17,41 +14,6 @@ $(() => {
     $bouton.on('click', () => axios.delete(`/api/service/${idService}`).then(() => window.location.reload()));
     $modaleSuppression.on('fermeModale', () => $bouton.off());
     $modaleSuppression.trigger('afficheModale');
-  });
-
-  const $modaleDuplication = $('.modale-duplication-service');
-  initialiseComportementModale($modaleDuplication);
-  $('.services').on('modaleDuplication', (_e, { idService, nomService }) => {
-    $('.nom-service', $modaleDuplication).text(nomService);
-
-    const selecteurFormulaire = '.modale-duplication-service form';
-    const $valider = $('.bouton-duplication-service', $modaleDuplication);
-    const $erreurServeur = $('.message-erreur-serveur', $modaleDuplication);
-
-    $(selecteurFormulaire).on('submit', (e) => {
-      e.preventDefault();
-      $erreurServeur.hide();
-      const nombreCopies = parseInt($('#nombre-copie').val(), 10) || 1;
-
-      tableauDeLongueur(nombreCopies).reduce((acc) => acc.then(
-        () => axios({ method: 'copy', url: `/api/service/${idService}` })
-      ), Promise.resolve())
-        .then(() => window.location.reload())
-        .catch((exc) => {
-          if (exc.response.status !== 424) return;
-
-          $erreurServeur.text(exc.response.data.message).show();
-        });
-    });
-
-    brancheValidation(selecteurFormulaire);
-    $(`${selecteurFormulaire} button[type = 'submit']`).on('click', () => declencheValidation(selecteurFormulaire));
-
-    $modaleDuplication.on('fermeModale', () => {
-      $erreurServeur.hide();
-      $valider.off();
-    });
-    $modaleDuplication.trigger('afficheModale');
   });
 
   const peupleServicesDans = (placeholder, donneesServices, idUtilisateur) => {
